+++ src/rest/routes/github-callback/index.test.ts
@@ -1,7 +1,5 @@
 import supertest from "supertest";
-import nock from "nock";
 import { getFrontendApp } from "~/src/app";
-import { envVars } from "config/env";
 
 describe("rest oauth router", () => {
 	let app;
@@ -12,22 +10,15 @@ describe("rest oauth router", () => {
 		describe("cloud", () => {
 			it("should exchange for github access token", async () => {
 				const code = "abcd";
-				const nockUrl = `/login/oauth/access_token?client_id=${envVars.GITHUB_CLIENT_ID}&client_secret=${envVars.GITHUB_CLIENT_SECRET}&code=${code}&state=`;
-				nock("https://github.com")
-					.get(nockUrl)
-					.matchHeader("accept", "application/json")
-					.matchHeader("content-type", "application/json")
-					.reply(200, {
-						access_token: "behold!",
-						refresh_token: "my-refresh-token"
-					});
+				const state = "qwer";
 
 				const resp = await supertest(app)
-					.get(`/rest/app/cloud/github-callback?code=${code}`)
+					.get(`/rest/app/cloud/github-callback?code=${code}&state=${state}`)
 					.expect("content-type", "text/html; charset=utf-8");
 
-				expect(resp.text).toEqual(expect.stringContaining("behold!"));
-				expect(resp.text).toEqual(expect.stringContaining("my-refresh-token"));
+				expect(resp.text).toEqual(expect.stringContaining(
+					`window.opener.postMessage({"code":"abcd","state":"qwer"}`
+				));
 
 			});
 		});
