+++ src/rest/routes/oauth/index.test.ts
@@ -1,4 +1,6 @@
 import supertest from "supertest";
+import nock from "nock";
+import { envVars } from "config/env";
 import { encodeSymmetric } from "atlassian-jwt";
 import { getFrontendApp } from "~/src/app";
 import { Installation } from "~/src/models/installation";
@@ -35,6 +37,32 @@ describe("rest oauth router", () => {
 				expect(resp.body).toEqual({
 					redirectUrl: expect.stringContaining("github-callback")
 				});
+			});
+		});
+	});
+	describe("exchange token", () => {
+		describe("cloud", () => {
+			it("should exchange for github access token", async () => {
+				const code = "abcd";
+				const nockUrl = `/login/oauth/access_token?client_id=${envVars.GITHUB_CLIENT_ID}&client_secret=${envVars.GITHUB_CLIENT_SECRET}&code=${code}&state=`;
+				nock("https://github.com")
+					.get(nockUrl)
+					.matchHeader("accept", "application/json")
+					.matchHeader("content-type", "application/json")
+					.reply(200, {
+						access_token: "behold!",
+						refresh_token: "my-refresh-token"
+					});
+
+				const resp = await supertest(app)
+					.post("/rest/app/cloud/oauth/exchangeToken")
+					.set("authorization", `${getToken()}`)
+					.send({ code })
+					.expect("content-type", "text/html; charset=utf-8");
+
+				expect(resp.text).toEqual(expect.stringContaining("behold!"));
+				expect(resp.text).toEqual(expect.stringContaining("my-refresh-token"));
+
 			});
 		});
 	});
